Saving songs with correct id3v2 tags

+++ Audiotica.Data.Collection/Model/Song.cs
@@ -31,7 +31,7 @@ namespace Audiotica.Data.Collection.Model
         //Artist prop is for the album (main), this one is specific to each song
         public string ArtistName { get; set; }
 
-        public long TrackNumber { get; set; }
+        public int TrackNumber { get; set; }
 
         public string AudioUrl { get; set; }
